% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvMF-package.R
\docType{package}
\name{rvMF-package}
\alias{rvMF-package}
\title{rvMF: Fast Generation of von Mises-Fisher Distributed Pseudo-Random Vectors}
\description{
Generates pseudo-random vectors that follow an arbitrary von Mises-Fisher distribution on a sphere. This method is fast and efficient when generating a large number of pseudo-random vectors. Functions to generate random variates and compute density for the distribution of an inner product between von Mises-Fisher random vector and its mean direction are also provided. Details are in Kang and Oh (2024) \doi{10.1007/s11222-024-10419-3}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/seungwoo-stat/rvMF}
  \item Report bugs at \url{https://github.com/seungwoo-stat/rvMF/issues}
}

}
\author{
\strong{Maintainer}: Seungwoo Kang \email{kangsw@skku.edu} (\href{https://orcid.org/0000-0001-8082-0794}{ORCID})

Authors:
\itemize{
  \item Hee-Seok Oh \email{heeseok@stats.snu.ac.kr}
}

}
\keyword{internal}
